// examples/uss-omaha_batch-4/uss-omaha/src/types/content.ts
export interface ExecutionPhotos {
  heading: string
  photos: ExecutionPhoto[]
}

export interface ExecutionPhoto {
  src: string
  caption: string
  year: string
}